LayoutTests/platform/chromium-cg-mac/svg/custom/use-nested-missing-target-removed-expected.txt

CONSOLE MESSAGE: line 7: Uncaught TypeError: Cannot set property 'id' of null
PASS
